backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.1
gunicorn>=21.2.0
redis>=5.0.1
celery>=5.3.6
django-unfold>=0.20.0
Pillow>=10.2.0
djangorestframework-simplejwt>=5.3.1
drf-spectacular>=0.27.1
django-filter>=24.1
python-dotenv>=1.0.1
channels>=4.0.0
channels-redis>=4.2.0
uvicorn[standard]>=0.27.1
django-import-export>=3.3.0
